factor out the inner workings of the indexes table so we can have separate regular and search ones
