+++ packages/compass-indexes/src/components/indexes-table/index.ts
@@ -1,0 +1,1 @@
+export * from './indexes-table';
