+++ packages/compass-indexes/src/components/regular-indexes-table/badge-with-icon-link.spec.tsx
