+++ packages/compass-indexes/src/components/regular-indexes-table/index-actions.spec.tsx
