+++ packages/compass-indexes/src/components/regular-indexes-table/size-field.tsx
