+++ packages/compass-indexes/src/components/regular-indexes-table/type-field.tsx
